+++ http/src/test/scala/com/twitter/finatra/http/tests/integration/routing/HttpServerAdminRouteTest.scala
@@ -1,6 +1,6 @@
 package com.twitter.finatra.http.tests.integration.routing
 
-import com.twitter.finagle.http.{RouteIndex, HttpMuxer, Request}
+import com.twitter.finagle.http._
 import com.twitter.finatra.http.routing.HttpRouter
 import com.twitter.finatra.http.EmbeddedHttpServer
 import com.twitter.finatra.http.{Controller, HttpServer}
@@ -180,6 +180,58 @@ class HttpServerAdminRouteTest extends WordSpecTest {
         intercept[java.lang.AssertionError] {
           server.start()
         }
+      } finally {
+        server.close()
+      }
+    }
+
+    "Similar paths but differing Http Methods" in {
+      val server = new EmbeddedHttpServer(
+        twitterServer = new HttpServer {
+          override protected def configureHttp(router: HttpRouter) = {
+            router.add(new Controller {
+
+              post("/foo/bar") { request: Request =>
+                response.ok.location("baz")
+              }
+
+              get("/foo/bar", admin = true) { request: Request =>
+                "baz"
+              }
+
+              post("/great/work", admin = true) { request: Request =>
+                "this is an admin POST route"
+              }
+
+              get("/great/work") { request: Request =>
+                "this is an external GET route"
+              }
+            })
+          }
+        })
+
+      try {
+        server.start()
+
+        server.adminHttpServerRoutes.map(_.path).contains("/foo/bar") should be(true) // a route with the path is there
+        server.adminHttpServerRoutes.exists(route => route.path == "/foo/bar" && route.method == Method.Get) should be(true) // however, we should also check the Method
+        server.adminHttpServerRoutes.exists(route => route.path == "/foo/bar" && route.method == Method.Post) should be(false) // in this case, the path with a Post does not exist
+
+        server.adminHttpServerRoutes.map(_.path).contains("/great/work") should be(true) // a route with the path is there
+        server.adminHttpServerRoutes.exists(route => route.path == "/great/work" && route.method == Method.Post) should be(true) // however, we should also check the Method
+        server.adminHttpServerRoutes.exists(route => route.path == "/great/work" && route.method == Method.Get) should be(false) // in this case, the path with a Get does not exist
+
+        server.httpPost(
+          "/foo/bar",
+          "",
+          andExpect = Status.Ok,
+          withLocation = "baz")
+
+        server.httpGet(
+          "/foo/bar",
+          andExpect = Status.Ok,
+          withBody = "baz")
+
       } finally {
         server.close()
       }